LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/block-local-documents-inheriting-none.https-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: undefined is not a constructor (evaluating 'new frame.contentWindow.ReportingObserver(
    rs => reports.push(...rs.map(r => r.toJSON()))
  )')

Harness Error (FAIL), message = Unhandled rejection: undefined is not a constructor (evaluating 'new frame.contentWindow.ReportingObserver(
    rs => reports.push(...rs.map(r => r.toJSON()))
  )')

FAIL Prevent local scheme documents from loading within a COEP: require-corp iframe if they inherit COEP: none promise_test: Unhandled rejection with value: object "TypeError: undefined is not a constructor (evaluating 'new frame.contentWindow.ReportingObserver(
    rs => reports.push(...rs.map(r => r.toJSON()))
  )')"
